backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
werkzeug==3.0.1
google-generativeai>=0.4.1,<0.5.0
pypdf==4.0.1
langchain==0.1.9
langchain-google-genai==0.0.11
python-multipart==0.0.9
trafilatura==1.6.1
beautifulsoup4==4.12.3
faiss-cpu==1.7.4
numpy>=1.24.0
lxml>=5.1.0
lxml_html_clean>=0.1.0
justext>=3.0.0
pymupdf4llm==0.0.24 